# ENH: Use ``PyObject_GetOptionalAttr``