Now that `PyObject_GetOptionalAttr` is public in 3.13 and semi-public via the capi-backports... Let's use it!

This is a massive speed up for code paths that run into this.  For full throttle gains, you need 3.12, because only in 3.12 were type lookups improved (I didn't test it with it).

For example:
```
o = object()  # some object that doesn't have __array__, or the other dunders

%timeit np.asarray(o)
```
will go from 800+ns to <200ns on my setup.  Things like `np.result_type(dtype, dtype)` (dtypes hit slow lookups for `__array_function__`) should be at 2-3 times faster on Python 3.12 (untested as I didn't feel like bumping my Python version).


Just to give another example, a Python object that implements a no-op `__array__` (i.e. returning an already wrapped array) goes from 575ns to 175ns.


---

Draft, because I had to base it on gh-27117 to avoid the conflict.