@seberg 2024-08-06
We should be using `GetOptionalAttr` in all places where an `AttributeError` is ignored.  But... we also should really almost never use `GetAttrStr` in code branches that may be relevant for performance, so I think doing just the protocols is a nice step.