@seberg 2024-08-07
Doesn't affect many inputs (but not few either, since a lot of objects have an `__array__` attribute), but just to show that it does show up e.g. in the array-coercion benchmarks:
```
| Change   | Before [d6021964] <main>   | After [fbd9a3b9] <use-optional-attr>   |   Ratio | Benchmark (Parameter)                                                            |
|----------|----------------------------|----------------------------------------|---------|----------------------------------------------------------------------------------|
| -        | 1.18±0.01μs                | 508±10ns                               |    0.43 | bench_array_coercion.ArrayCoercionSmall.time_array_all_kwargs(range(0, 3))       |
| -        | 1.08±0.01μs                | 426±3ns                                |    0.4  | bench_array_coercion.ArrayCoercionSmall.time_asanyarray_dtype(range(0, 3))       |
| -        | 1.07±0.01μs                | 429±3ns                                |    0.4  | bench_array_coercion.ArrayCoercionSmall.time_asarray_dtype(range(0, 3))          |
| -        | 1.08±0.01μs                | 421±20ns                               |    0.39 | bench_array_coercion.ArrayCoercionSmall.time_array_subok(range(0, 3))            |
| -        | 1.09±0.01μs                | 417±4ns                                |    0.38 | bench_array_coercion.ArrayCoercionSmall.time_array_dtype_not_kwargs(range(0, 3)) |
| -        | 1.08±0μs                   | 412±3ns                                |    0.38 | bench_array_coercion.ArrayCoercionSmall.time_array_no_copy(range(0, 3))          |
| -        | 1.06±0μs                   | 396±5ns                                |    0.37 | bench_array_coercion.ArrayCoercionSmall.time_array(range(0, 3))                  |
| -        | 1.03±0μs                   | 386±0.2ns                              |    0.37 | bench_array_coercion.ArrayCoercionSmall.time_asanyarray(range(0, 3))             |
| -        | 1.03±0.01μs                | 383±3ns                                |    0.37 | bench_array_coercion.ArrayCoercionSmall.time_ascontiguousarray(range(0, 3))      |
| -        | 1.05±0.02μs                | 381±3ns                                |    0.36 | bench_array_coercion.ArrayCoercionSmall.time_asarray(range(0, 3))                |

SOME BENCHMARKS HAVE CHANGED SIGNIFICANTLY.
PERFORMANCE INCREASED.
```