@charris 2024-08-10
I'd like to backport the `pythoncapi-compat` submodule to keep the 2.1 branch in sync. Any problem with that?